+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 1.x branch
 ## 1.8 branch
+### 1.8.3
+* `url_to_absolute()` is deprecated and will be removed in a later release;
+* `slug()` is deprecated and will be removed in a later release.
+
 ### 1.8.2
 * added tests for PHP 8.2;
 * `BodyParser` is deprecated and will be removed in a later release;